+++ tests/testthat/test-ord_explore.R
@@ -1,0 +1,8 @@
+library(shinytest)
+
+test_that("ord_explore() works", {
+  testthat::skip_on_cran()
+  testthat::skip_on_bioc()
+
+  expect_pass(testApp(appDir = "apps/ord_explore/", compareImages = FALSE))
+})